pyproject.toml
[project]
name = "blog-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "boto3>=1.35.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.9.0",
]

[dependency-groups]
dev = [
    "uvicorn[standard]>=0.30.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "moto[dynamodb,s3]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/admin", "src/shared"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-m 'not integration'"
markers = [
    "integration: requires a live server + DynamoDB Local (run via scripts/integration_test.sh)",
]
